+++ RELEASES.md
@@ -1,5 +1,11 @@
 # Release notes
 
+### 0.0.7 / Mon 10 Jun 2018
+
+* Adding `<f-piano>` onscreen keyboard component
+* Adding `chords()` and `chord()` music utility functions
+* Adding `intersection()` array utility function
+
 ### 0.0.6 / Mon 3 Jun 2018
 
 #### New features